refactor: rename AppSidePanelView.jsx to index.jsx
